Reorganized packages a bit.
