+++ src/main/java/technology/sola/json/parser/Parser.java
@@ -1,9 +1,9 @@
 package technology.sola.json.parser;
 
 import technology.sola.json.exception.InvalidSyntaxException;
-import technology.sola.json.token.Token;
-import technology.sola.json.token.TokenType;
-import technology.sola.json.token.Tokenizer;
+import technology.sola.json.tokenizer.Token;
+import technology.sola.json.tokenizer.TokenType;
+import technology.sola.json.tokenizer.Tokenizer;
 
 import java.util.ArrayList;
 import java.util.List;
