+++ src/main/java/technology/sola/json/tokenizer/TokenType.java
@@ -1,4 +1,4 @@
-package technology.sola.json.token;
+package technology.sola.json.tokenizer;
 
 public enum TokenType {
   COLON,
