+++ src/test/java/technology/sola/json/TestPojo.java
@@ -1,7 +1,9 @@
 package technology.sola.json;
 
-record TestPojo(int value, String value2) {
-  static JsonMapper<TestPojo> JSON_MAPPER = new JsonMapper<>() {
+import technology.sola.json.mapper.JsonMapper;
+
+public record TestPojo(int value, String value2) {
+  public static JsonMapper<TestPojo> JSON_MAPPER = new JsonMapper<>() {
     @Override
     public Class<TestPojo> getObjectClass() {
       return TestPojo.class;
