+++ src/test/java/technology/sola/json/parser/ParserTest.java
@@ -3,8 +3,8 @@ package technology.sola.json.parser;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import technology.sola.json.token.TokenType;
-import technology.sola.json.token.Tokenizer;
+import technology.sola.json.tokenizer.TokenType;
+import technology.sola.json.tokenizer.Tokenizer;
 
 class ParserTest {
   @Nested
